requirements.txt
fastapi==0.108.0
uvicorn==0.24.0
python-multipart==0.0.6
pymupdf==1.23.5
python-docx==1.0.1
google-generativeai==0.3.1
python-dotenv==1.0.0
pydantic==2.4.2